requirements.txt
asgiref==3.6.0  
Django==4.1.5  
django-cors-headers==3.13.0 
djangorestframework==3.14.0 
pip==22.3.1 
psycopg2==2.9.5  
pytz==2022.7 
setuptools==65.6.3 
sqlparse==0.4.3  
tzdata==2022.7 
wheel==0.38.4 